+++ controllers/categoriesController.js
@@ -9,7 +9,7 @@ exports.getAllCategories = (request, response, next) => {
   })
   .catch(error => {
     response.status(500).json({
-      error:error
+      "status":"Error, Could not fetch record!"
     });
   });
 };
@@ -32,7 +32,7 @@ exports.createCategory = (request, res, next) => {
   })
   .catch((error) => {
     res.status(500).json({
-      error: error
+      "status":"Error, Could not save record!"
     });
   });
 };
@@ -48,7 +48,7 @@ exports.getOneCategory = (request, response, next) => {
   })
   .catch((error) => {
     response.status(500).json({
-      error:error
+      "status":"Error, Could not fetch record!"
     });
   });
 };
@@ -75,23 +75,36 @@ exports.modifyCategory = (request, response, next) => {
   })
   .catch((error) => {
     response.status(500).json({
-      error: error
+      "status":"Error, Could not update record!"
     });
   });
 };
 
 
 exports.deleteCategory = (request, response, next) => {
-  dbconn.query('DELETE FROM categories WHERE id = $1', [request.params.id])
+  dbconn.query('SELECT id, userid FROM categories WHERE id = $1', [request.params.id])
   .then((data) => {
-    response.status(201).json({
-      "status":"success",
-      "data":data.rows[0]
-    });
-  })
-  .catch((error) => {
+    if ((data.rows[0].userid != request.body.currUserId) && (Number(request.body.currUserRole) !== 1)) {
+      response.status(201).json({
+        "status":"Access denied!",
+      });
+    } else { 
+      dbconn.query('DELETE FROM categories WHERE id = $1', [request.params.id])
+      .then((data) => {
+        response.status(201).json({
+          "status":"success",
+          "data":data.rows[0]
+        });
+      })
+      .catch((error) => {
+        response.status(500).json({
+          "status":"Error, Could not delete record!"
+        });
+      });
+    }
+  }).catch((error) => {
     response.status(500).json({
-      error:error
+      "status":"Error, Could not Resolve Item to delete!"
     });
   });
 };
